@@ -35,10 +35,10 @@
     "open-color": "^1.8.0",
     "perfect-freehand": "0.4.7",
     "postcss": "^8.3.5",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.18.0",
     "react-states": "^6.2.3",
     "roughjs": "^4.4.1",
     "sass": "^1.32.8",
@@ -51,8 +51,8 @@
     "@babel/preset-typescript": "^7.13.0",
     "@types/jest": "^26.0.20",
     "@types/node": "^14.14.34",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^18.2.34",
+    "@types/react-dom": "^18.2.14",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-test-renderer": "^17.0.1",
     "@vitejs/plugin-react-refresh": "^1.1.0",
