@@ -45,9 +45,7 @@ export const ExcalidrawCanvas = React.memo(
 
     useEffect(() => {
       import("@excalidraw/excalidraw").then((comp) => {
-        // When building it, it exports on default.default
-        // @ts-ignore
-        setComp(comp.default?.default ?? comp.default);
+        setComp(comp.Excalidraw);
       });
     }, [Comp]);
 
@@ -73,11 +71,11 @@ export const ExcalidrawCanvas = React.memo(
                 if (changedData) {
                   excalidrawRef.current.updateScene(changedData);
                 }
-              }
+              },
             );
           }
         }),
-      []
+      [],
     );
 
     return (
@@ -92,5 +90,5 @@ export const ExcalidrawCanvas = React.memo(
         ) : null}
       </div>
     );
-  }
+  },
 );
